@DharanCosta [Task: Validanting checklist from @GustavoBoaz, all requests complete]

+++ src/main/java/org/blogpessoal/blogpessoal/controller/PostagemController.java
@@ -1,6 +1,9 @@
 package org.blogpessoal.blogpessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.blogpessoal.blogpessoal.model.Postagem;
 import org.blogpessoal.blogpessoal.repository.PostagemRepository;
@@ -15,13 +18,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 // Configurações do Controller para receber requisições \/
 @RestController 
 @RequestMapping("/postagens") 
-@CrossOrigin("*")
-
+@CrossOrigin(origins = "*", allowedHeaders ="*")
 public class PostagemController {
 
 	@Autowired  // Autowired é a gestão de dependencias do Spring - para acessar os repositórios
@@ -49,19 +53,25 @@ public class PostagemController {
 	
 	//Método Post 
 	@PostMapping
-	public ResponseEntity<Postagem> post (@RequestBody Postagem postagem){
+	public ResponseEntity<Postagem> post (@Valid @RequestBody Postagem postagem){
 		return ResponseEntity.status(HttpStatus.CREATED).body(repositoty.save(postagem));
 	}
 	
 	//Método Put
 	@PutMapping
-	public ResponseEntity<Postagem> put (@RequestBody Postagem postagem){
+	public ResponseEntity<Postagem> put (@Valid @RequestBody Postagem postagem){
 		return ResponseEntity.status(HttpStatus.OK).body(repositoty.save(postagem));
 	}
 	
 	//Método Delete
+	@ResponseStatus (HttpStatus.NO_CONTENT)
 	@DeleteMapping("/{id}")
 	public void delete(@PathVariable long id){
+		Optional<Postagem> post = repositoty.findById(id);
+
+		if(post.isEmpty())
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
 		repositoty.deleteById(id);
 	}
 }
