+++ src/main/java/org/blogpessoal/blogpessoal/controller/TemaController.java
@@ -1,6 +1,9 @@
 package org.blogpessoal.blogpessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.blogpessoal.blogpessoal.model.Tema;
 import org.blogpessoal.blogpessoal.repository.TemaRepository;
@@ -15,7 +18,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 
 @RestController
@@ -43,18 +48,26 @@ public class TemaController {
 		}
 		
 		@PostMapping
-		public ResponseEntity<Tema> post(@RequestBody Tema tema){
+		public ResponseEntity<Tema> post(@Valid @RequestBody Tema tema){
 			return ResponseEntity.status(HttpStatus.CREATED)
 					.body(repository.save(tema));
 		}
 		
 		@PutMapping
-		public ResponseEntity<Tema> put (@RequestBody Tema tema){
-			return ResponseEntity.ok(repository.save(tema));
+		public ResponseEntity<Tema> put (@Valid @RequestBody Tema tema){
+			return repository.findById(tema.getId())
+					.map(resp -> ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema)))
+					.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 		}
 		
+		@ResponseStatus(HttpStatus.NO_CONTENT)
 		@DeleteMapping("/{id}")
 		public void delete(@PathVariable long id) {
+			Optional<Tema> tema = repository.findById(id);
+
+			if(tema.isEmpty())
+				throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
 			repository.deleteById(id);
 		}
